fix bag with forgot password and activate account

+++ auth/internal/lib/parseJWT/parse_jwt.go
@@ -13,7 +13,6 @@ func ParseToken(tokenString string) (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-
 	if !token.Valid {
 		return 0, errors.New("invalid token")
 	}
@@ -22,7 +21,6 @@ func ParseToken(tokenString string) (int64, error) {
 	if !ok {
 		return 0, errors.New("invalid claims")
 	}
-
 	expInt, ok := claims["exp"].(float64)
 	if !ok {
 		return 0, errors.New("invalid exp")
@@ -31,9 +29,9 @@ func ParseToken(tokenString string) (int64, error) {
 	if time.Now().After(time.Unix(int64(expInt), 0)) {
 		return 0, errors.New("expired")
 	}
-	userID, ok := claims["uid"].(int64)
+	userID, ok := claims["uid"].(float64)
 	if !ok {
 		return 0, errors.New("invalid userID")
 	}
-	return userID, nil
+	return int64(userID), nil
 }
